chore: move testing.ts into tests/
